third_party/blink/web_tests/http/tests/devtools/a11y-axe-core/settings/emulated-device-a11y-test-expected.txt

Test error message in the settings tool Emulated Device pane

Running: testNewDeviceError
Invalidating the device pixel ratio
Error message: Device pixel ratio must be a number or blank.
